+++ src/pdfnaut/objects/annotations/base.py
@@ -5,7 +5,7 @@ from typing import Literal, cast, get_args, overload
 from typing_extensions import Self
 
 from pdfnaut.common.dictmodels import dictmodel, field
-from pdfnaut.cos.objects.base import PdfName, PdfReference, is_null_like
+from pdfnaut.cos.objects.base import PdfName, PdfReference
 from pdfnaut.cos.objects.containers import PdfArray, PdfDictionary
 from pdfnaut.cos.parser import PdfParser
 from pdfnaut.exceptions import PdfParseError, PdfWriteError
@@ -131,15 +131,10 @@ class Annotation(PdfDictionary):
     See ISO 32000-2:2020 § 14.9.2 "Natural language specification" for details.
     """
 
-    flags: AnnotationFlags = field("F", default=AnnotationFlags.NULL.value)
+    flags: AnnotationFlags = field("F", default=AnnotationFlags.NULL)
     """Flags specifying various characteristics of the annotation."""
 
-    color: list[float] | None = field(
-        "C",
-        default=None,
-        encoder=lambda lst: PdfArray(lst) if lst is not None else None,
-        decoder=lambda arr: list(arr) if not is_null_like(arr) else None,
-    )
+    color: list[float] | None = field("C", default=None, encoder=PdfArray, decoder=list)
     """An array of 0 to 4 numbers in the range 0.0 to 1.0, representing a color used
     for the following purposes:
 
@@ -256,12 +251,7 @@ class AnnotationBorderStyle(PdfDictionary):
     )
     """The border style."""
 
-    dash_pattern: list[int] | None = field(
-        "D",
-        default=None,
-        encoder=lambda lst: PdfArray(lst) if lst is not None else None,
-        decoder=lambda arr: list(arr) if not is_null_like(arr) else None,
-    )
+    dash_pattern: list[int] | None = field("D", default=None, encoder=PdfArray, decoder=list)
     """The dash pattern that will be used for the border if the style specified
     is dashed. The array consists of alternating dashes and gaps. The dash phase
     is not specified and is assumed to be zero."""