+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This is the changelog for pdfnaut. Versions follow the scheme specified in the [
 ### Fixes
 
 - Fix wrong type hint of the `ids` parameter of `StandardSecurityHandler`.
+- Fix `LineAnnotation.line_ending_color` and `RectangleAnnotation.interior_color` not being usable due to missing field encoder/decoder.
 
 ## [0.13.0] (2026-08-07)